+++ account-kit/infra/src/chains.ts
@@ -8,6 +8,8 @@ import {
   base as vbase,
   baseGoerli as vbaseg,
   baseSepolia as vbases,
+  bsc as vbsc,
+  bscTestnet as vbsct,
   fraxtal as vfrax,
   goerli as vgo,
   mainnet as vmain,
@@ -172,6 +174,26 @@ export const baseSepolia: Chain = {
     ...vbases.rpcUrls,
     alchemy: {
       http: ["https://base-sepolia.g.alchemy.com/v2"],
+    },
+  },
+};
+
+export const bsc: Chain = {
+  ...vbsc,
+  rpcUrls: {
+    ...vbsc.rpcUrls,
+    alchemy: {
+      http: ["https://bnb-mainnet.g.alchemy.com/v2"],
+    },
+  },
+};
+
+export const bscTestnet: Chain = {
+  ...vbsct,
+  rpcUrls: {
+    ...vbsct.rpcUrls,
+    alchemy: {
+      http: ["https://bnb-testnet.g.alchemy.com/v2"],
     },
   },
 };
@@ -689,6 +711,30 @@ export const celoMainnet: Chain = defineChain({
   testnet: false,
 });
 
+export const celoSepolia: Chain = defineChain({
+  id: 11142220,
+  name: "Celo Sepolia",
+  nativeCurrency: { name: "Celo native asset", symbol: "CELO", decimals: 18 },
+  rpcUrls: {
+    default: {
+      http: ["https://celo-sepolia.g.alchemy.com/v2"],
+    },
+    public: {
+      http: ["https://celo-sepolia.g.alchemy.com/v2"],
+    },
+    alchemy: {
+      http: ["https://celo-sepolia.g.alchemy.com/v2"],
+    },
+  },
+  blockExplorers: {
+    default: {
+      name: "Block Explorer",
+      url: "https://celo-sepolia.blockscout.com/",
+    },
+  },
+  testnet: true,
+});
+
 export const teaSepolia: Chain = defineChain({
   id: 10218,
   name: "Tea Sepolia",
@@ -756,6 +802,30 @@ export const bobaMainnet: Chain = defineChain({
     default: {
       name: "Block Explorer",
       url: "https://bobascan.com/",
+    },
+  },
+  testnet: false,
+});
+
+export const edgeMainnet: Chain = defineChain({
+  id: 3343,
+  name: "Edge Mainnet",
+  nativeCurrency: { name: "eth", symbol: "eth", decimals: 18 },
+  rpcUrls: {
+    default: {
+      http: ["https://edge-mainnet.g.alchemy.com/v2"],
+    },
+    public: {
+      http: ["https://edge-mainnet.g.alchemy.com/v2"],
+    },
+    alchemy: {
+      http: ["https://edge-mainnet.g.alchemy.com/v2"],
+    },
+  },
+  blockExplorers: {
+    default: {
+      name: "Block Explorer",
+      url: "https://edge-mainnet.blockscout.com/",
     },
   },
   testnet: false,